move custom keybinds code to module
